Remove goldfish/hello.c
Main is multiply-defined at go compile time if hello.c is present. The
functionality of hello.c is duplicated in main.go so we don't need
it. It seems that the golang compilation process is aware of all c files
in the directory, the way this project is set up.